// frontend/src/client/types.gen.ts
export type WatchlistCreate = {
    name: string;
    description?: (string | null);
    is_default?: boolean;
    is_active?: boolean;
};

export type WatchlistItemCreate = {
    notes?: (string | null);
    stock_id: string;
};

export type WatchlistItemPublic = {
    notes?: (string | null);
    id: string;
    watchlist_id: string;
    stock_id: string;
    added_at: string;
};

export type WatchlistPublic = {
    name: string;
    description?: (string | null);
    is_default?: boolean;
    is_active?: boolean;
    id: string;
    user_id: string;
    created_at: string;
    updated_at: string;
};

export type WatchlistUpdate = {
    name?: (string | null);
    description?: (string | null);
    is_default?: (boolean | null);
    is_active?: (boolean | null);
};